Moved countbufferusers
