+++ src/ExtractHWBuffers.cpp
@@ -106,124 +106,6 @@ std::vector<MergedDimSize> create_hwbuffer_sizes(std::vector<int> logical_size,
 namespace {
 
 
-/* 
- * Used to calculate input and output block size. 
- * Also records the (writer and) reader loopnest.
- */  
-class CountBufferUsers : public IRVisitor {
-  using IRVisitor::visit;
-  string var;
-  Scope<Expr> scope;
-  For *current_for;
-  Stmt full_stmt;
-
-  // keep track of lets to later replace variables with constants
-  void visit(const LetStmt *op) override {
-      ScopedBinding<Expr> bind(scope, op->name, simplify(expand_expr(op->value, scope)));
-      IRVisitor::visit(op);
-  }
-
-  void visit(const ProducerConsumer *op) override {
-    if (!op->is_producer) {
-      if (provide_at_level(op->body, var)) {
-        auto box_write = box_provided(op->body, var);
-        auto interval = box_write;
-        input_block_box = vector<Expr>(interval.size());
-        for (size_t dim=0; dim<interval.size(); ++dim) {
-          Expr port_expr = simplify(expand_expr(interval[dim].max - interval[dim].min + 1, scope));
-          Expr lower_expr = find_constant_bound(port_expr, Direction::Lower);
-          Expr upper_expr = find_constant_bound(port_expr, Direction::Upper);
-          input_block_box[dim] = is_undef(lower_expr) ? port_expr : lower_expr;
-        }
-
-      }
-
-      // look at the readers (reader ports, read address gen)
-      if (call_at_level(op->body, var)) {
-        auto box_read = box_required(op->body, var);
-        auto interval = box_read;
-
-        output_block_box = vector<Expr>(interval.size());
-        for (size_t dim=0; dim<interval.size(); ++dim) {
-          Expr port_expr = simplify(expand_expr(interval[dim].max - interval[dim].min + 1, scope));
-          Expr lower_expr = find_constant_bound(port_expr, Direction::Lower);
-          Expr upper_expr = find_constant_bound(port_expr, Direction::Upper);
-          output_block_box[dim] = is_undef(lower_expr) ? port_expr : lower_expr;
-        }
-      }
-    }
-    IRVisitor::visit(op);
-  }
-
-  void visit(const For *op) override {
-    // add this for loop to read address loop
-    For *previous_for = current_for;
-    Stmt for_stmt = For::make(op->name, op->min, op->extent, op->for_type, op->device_api, op->body);
-    if (current_for == nullptr) {
-      full_stmt = for_stmt;
-
-      auto box_read = box_required(op->body, var);
-      auto interval = box_read;
-      for (size_t dim=0; dim<interval.size(); ++dim) {
-        auto assertstmt = AssertStmt::make(var + "_dim" + std::to_string(dim), simplify(expand_expr(interval[dim].min, scope)));
-      }
-    } else if (!is_parallelized(op)) {
-      current_for->body = for_stmt;
-    }
-    current_for = const_cast<For *>(for_stmt.as<For>());
-
-    IRVisitor::visit(op);
-
-    // look at the writers (writer ports)
-    if (provide_at_level(op->body, var) && !is_parallelized(op)) {
-      auto box_write = box_provided(op->body, var);
-      auto interval = box_write;
-      input_block_box = vector<Expr>(interval.size());
-      for (size_t dim=0; dim<interval.size(); ++dim) {
-        Expr port_expr = simplify(expand_expr(interval[dim].max - interval[dim].min + 1, scope));
-        Expr lower_expr = find_constant_bound(port_expr, Direction::Lower);
-        Expr upper_expr = find_constant_bound(port_expr, Direction::Upper);
-        input_block_box[dim] = is_undef(lower_expr) ? port_expr : lower_expr;
-      }
-
-    }
-
-    // look at the readers (reader ports, read address gen)
-    if (call_at_level(op->body, var) && !is_parallelized(op)) {
-      auto box_read = box_required(op->body, var);
-      auto interval = box_read;
-
-      vector<Stmt> stmts;
-      output_block_box = vector<Expr>(interval.size());
-      for (size_t dim=0; dim<interval.size(); ++dim) {
-        Expr port_expr = simplify(expand_expr(interval[dim].max - interval[dim].min + 1, scope));
-        Expr lower_expr = find_constant_bound(port_expr, Direction::Lower);
-        Expr upper_expr = find_constant_bound(port_expr, Direction::Upper);
-        output_block_box[dim] = is_undef(lower_expr) ? port_expr : lower_expr;
-        stmts.push_back(AssertStmt::make(var + "_dim" + std::to_string(dim), simplify(expand_expr(interval[dim].min, scope))));
-      }
-
-      const vector<Stmt> &const_stmts = stmts;
-      current_for->body = Block::make(const_stmts);
-      reader_loopnest = simplify(expand_expr(full_stmt, scope));
-    }
-
-
-    // remove for loop to read address loop
-    current_for = previous_for;
-    
-  }
-  
-public:
-  vector<Expr> output_block_box;
-  vector<Expr> input_block_box;
-  vector<std::string> for_loops;
-  Stmt reader_loopnest;
-
-  CountBufferUsers(string v) :
-    var(v), current_for(nullptr){}
-};
-
 class FindVarStride : public IRVisitor {
   string varname;
   string loopname;
